Texture sampling and fragment shader support for Spheres (ellipsoids) and Cubes (rectangular prisms) (#139)
* Implemented sphere texture projection

* Modified sphere

* Shader support for sphere and changed box a bit, still need more work

* Finished cube projection... for now...

* Rebranding and version bump to 1.1.7

---------

+++ RenderSharp/Render2d/Renderer2d/Renderer.cs
@@ -263,8 +263,8 @@ namespace RenderSharp.Render2d
                 //: Transforms.WorldToBgTexture2(worldLoc, scene.BgTexture.Size, scene.BgTextureWorldSize);
             Scene.BgCoordShader(bgTextureInd, out bgTextureInd, Resolution, scene.Time);
             RGBA outColor = scene.BgTexture[
-                Operations.Mod(bgTextureInd.X, scene.BgTexture.Width),
-                Operations.Mod(bgTextureInd.Y, scene.BgTexture.Height)];
+                (int)Operations.Mod(bgTextureInd.X, scene.BgTexture.Width),
+                (int)Operations.Mod(bgTextureInd.Y, scene.BgTexture.Height)];
 
             FRGBA fOut = outColor;
             Scene.BgFragShader(fOut, out fOut, bgTextureInd, scene.BgTexture.Size, scene.Time);
